Debugging libdb2.so issue
