+++ Frends.IBMDB2.ExecuteQuery/Frends.IBMDB2.ExecuteQuery/Frends.IBMDB2.ExecuteQuery.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	  <TargetFramework>net6.0</TargetFramework>
+	  <TargetFramework>net6</TargetFramework>
 	  <LangVersion>Latest</LangVersion>
 	  <Version>1.0.0</Version>
 	  <Authors>Frends</Authors>
@@ -29,7 +29,7 @@
 	</AssemblyAttribute>
   </ItemGroup>
 
-  <ItemGroup>
+  <ItemGroup>    
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
     <PackageReference Include="System.ComponentModel.Annotations" Version="4.7.0" />
@@ -37,11 +37,11 @@
   </ItemGroup>
 
 	<ItemGroup Condition="'$(OS)' == 'WINDOWS_NT'">
-		<PackageReference Include="Net.IBM.Data.Db2" Version="6.0.0.400" />
+		<PackageReference Include="IBM.Data.DB2.Core" Version="3.1.0.600" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(OS)' == 'Unix'">
-		<PackageReference Include="Net.IBM.Data.Db2-lnx" Version="6.0.0.400" />
+		<PackageReference Include="IBM.Data.DB2.Core-lnx" Version="3.1.0.500" />
 	</ItemGroup>
 	
 	<ItemGroup>
